requirements.txt
streamlit
langchain
ipykernel
python-dotenv
langchain_community
jq
jsonlines
langchain-text-splitters
chromadb
sentence_transformers
langchain_huggingface
langchain_chroma
langchain-groq
fastapi
uvicorn
langserve